//Update 배경스크롤 완전구현 몬스터 움직임직을 프레임타임으로 구현 원형충돌체크로 변경
git-svn-id: https://dev.naver.com/svn/eotktjdrkf@35 29f03ed2-db14-0510-ae99-898e00d8f4b7

+++ FrameWork_03/app/src/main/java/teamwarpcbstuido/IESEKI/game/BackGround.java
@@ -47,7 +47,7 @@ public class BackGround extends GraphicObject{
 
     }
 
-    public void onUpdate(long GameTime)
+    public void onUpdate(float GameTime)
     {
 
         m_background_01.offset(0, -DPI[Y]/8);
@@ -64,8 +64,8 @@ public class BackGround extends GraphicObject{
 
     @Override
     public void Draw(Canvas canvas){
-    canvas.drawBitmap(m_layer1, null, m_background_01, null);
-    canvas.drawBitmap(m_layer2, null, m_background_02, null);
+        canvas.drawBitmap(m_layer1, null, m_background_01, null);
+        canvas.drawBitmap(m_layer2, null, m_background_02, null);
     }
 
     public void onDraw(Canvas canvas) {
